receiver/requirements.txt
ndi-python
numpy
opencv-python
pillow
